+++ collectfasta/tests/utils.py
@@ -1,3 +1,4 @@
+from concurrent.futures import ThreadPoolExecutor
 import functools
 import os
 import pathlib
@@ -20,9 +21,27 @@ from collectfasta import settings
 
 live_test = pytest.mark.live_test
 
+speed_test_mark = pytest.mark.speed_test
+speed_test_option = pytest.mark.skipif("not config.getoption('speedtest')")
+speed_test = lambda func: speed_test_mark(speed_test_option(func))
+
 static_dir: Final = pathlib.Path(django_settings.STATICFILES_DIRS[0])
 
 F = TypeVar("F", bound=Callable[..., Any])
+
+
+def make_100_files():
+    with ThreadPoolExecutor(max_workers=5) as executor:
+        for _ in range(100):
+            executor.submit(create_big_static_file)
+        executor.shutdown(wait=True)
+
+
+def make_1000_files():
+    with ThreadPoolExecutor(max_workers=10) as executor:
+        for _ in range(1000):
+            executor.submit(create_static_file)
+        executor.shutdown(wait=True)
 
 
 def get_fake_client():
@@ -88,6 +107,13 @@ def create_static_file() -> pathlib.Path:
     """Write random characters to a file in the static directory."""
     path = static_dir / f"{uuid.uuid4().hex}.txt"
     path.write_text("".join(chr(random.randint(0, 64)) for _ in range(500)))
+    return path
+
+
+def create_big_static_file() -> pathlib.Path:
+    """Write random characters to a file in the static directory."""
+    path = static_dir / f"{uuid.uuid4().hex}.txt"
+    path.write_text("".join(chr(random.randint(0, 64)) for _ in range(100000)))
     return path